+++ activerecord/test/cases/scoping/relation_scoping_test.rb
@@ -20,11 +20,21 @@ class RelationScopingTest < ActiveRecord::TestCase
   end
 
   def test_reverse_order_with_multiple_arel_nodes
-    assert_equal Developer.order("id DESC").order("name DESC").to_a.reverse, Developer.order(Developer.arel_table[:id].desc).order(Developer.arel_table[:name].desc).reverse_order
+    assert_deprecated do
+      assert_equal Developer.order("id DESC").order("name DESC").to_a.reverse, Developer.order(Developer.arel_table[:id].desc).order(Developer.arel_table[:name].desc).reverse_order
+    end
+
+    assert_equal Developer.order("id DESC").order.append("name DESC").to_a.reverse, Developer.order(Developer.arel_table[:id].desc).order.append(Developer.arel_table[:name].desc).reverse_order
+    assert_equal Developer.order("id DESC").order.prepend("name DESC").to_a.reverse, Developer.order(Developer.arel_table[:id].desc).order.prepend(Developer.arel_table[:name].desc).reverse_order
   end
 
   def test_reverse_order_with_arel_nodes_and_strings
-    assert_equal Developer.order("id DESC").order("name DESC").to_a.reverse, Developer.order("id DESC").order(Developer.arel_table[:name].desc).reverse_order
+    assert_deprecated do
+      assert_equal Developer.order("id DESC").order("name DESC").to_a.reverse, Developer.order("id DESC").order(Developer.arel_table[:name].desc).reverse_order
+    end
+
+    assert_equal Developer.order("id DESC").order.append("name DESC").to_a.reverse, Developer.order("id DESC").order.append(Developer.arel_table[:name].desc).reverse_order
+    assert_equal Developer.order("id DESC").order.prepend("name DESC").to_a.reverse, Developer.order("id DESC").order.prepend(Developer.arel_table[:name].desc).reverse_order
   end
 
   def test_double_reverse_order_produces_original_order